Crate ibc_types_core_connection

IBC connection-related types.

Modules

events

Types for ABCI Events that inform relayers about IBC client events.

msgs

Message definitions for the connection handshake datagrams.

Structs

ChainId

This type is subject to future changes.

ClientPaths

ClientPaths contains a list of all of the connection paths that are associated with a specific client

ConnectionEnd

ConnectionId

Counterparty

IdentifiedConnectionEnd

Version

Stores the identifier and the features supported by a version

Enums

ConnectionError

State