ibc_types_core_connection
Module msgs

Message definitions for the connection handshake datagrams.

We define each of the four messages in the connection handshake protocol as a struct. Each such message comprises the same fields as the datagrams defined in ICS3 English spec: https://github.com/cosmos/ibc/tree/master/spec/core/ics-003-connection-semantics.

One departure from ICS3 is that we abstract the three counterparty fields (connection id, prefix, and client id) into a single field of type Counterparty; this applies to messages MsgConnectionOpenInit and MsgConnectionOpenTry. One other difference with regards to abstraction is that all proof-related attributes in a message are encapsulated in Proofs type.

Another difference to ICS3 specs is that each message comprises an additional field called signer which is specific to Cosmos-SDK.

Structs

MsgConnectionOpenAck

Per our convention, this message is sent to chain A. The handler will check proofs of chain B.

MsgConnectionOpenConfirm

Per our convention, this message is sent to chain B. The handler will check proofs of chain A.

MsgConnectionOpenInit

Per our convention, this message is sent to chain A. The handler will check proofs of chain B.

MsgConnectionOpenTry

Per our convention, this message is sent to chain B. The handler will check proofs of chain A.

Enums

ConnectionMsg

Enumeration of all possible messages that the ICS3 protocol processes.